Adding a queue column family to serve as an index
Iterating over all schedule ids to determine which rows have new data
that needs to be processed for calculating aggregrates would be really
slow; so, this commit introduces a new column family,
metrics_work_queue. This column family has its columns ordered by
time, and provides a fast, efficient way to find those schedules that
have data that need to be processed. Tests have been updated to verify
that the queue is updated.