More refactoring to support aggregating 6 hr and 24 hr data
This commit adds more support for aggregating data int the various
buckets. The new version of calculateAggregates takes the to/from column
family and the interval and minutes. Logic has been added to make sure
that the appropriate queue is updated as well.

A check has been added to make sure that the current time slice has
passed before aggregating the data for that slice.

Lastly, unit tests are undergoing substantial changes as I try to figure
out the most effective way to get good test coverage in place.