Fixing logic around how timestamps where getting set
I was previously using the collectionTime property of the measurement
report to set the timestamp of the metric data columns getting inserted.
That property though is not a timestamp but rather a duration for how
long the collection run took. The timestamps which getting rounded down
to the starting hour of the appropriate interval (except for raw data)
use the timestamps of the data themselves.