Adding a bundle for provisioning a host with Cassandra nodes
This commit updates the build to generate a bundle file that will
deploy and configure a Cassandra node to a specified host. Right now
this is intended for development environments, particularly for running
a cluster on a single host. The bundle will be packaged up with the
server plugin so that the provisioning can be exposed through a plugin
operation.