Adding logic to bundle recipe to execute schema creation script
The dbsetup.script file is a CLI script that creates the RHQ schema. It
is only run on the first node that is installed. This commit also
parameterizes another couple properties in cassandra.yaml. The
properties specify the host name which is normally resovled as
localhost. We need to override this for a local, development cluster
where all nodes run on a single machine.