Changing packaging of cassandra bundle from zip to jar
Whenever I redeployed the server plugin there were IO exceptions that
occurred when trying to read the cassandra bundle. Restarting my RHQ
server resolved the problems. Not sure what was casusing the IO
exceptions, but changing the bundle to a jar file instead of a zip file
resolved the problem so that a server restart is not required.