refactoring build to use cassandra-maven-plugin for testing
This commit adds support for the cassandra-maven-plugin for
unit/integration tests that require a running cassandra instance or
cassandra cluster.

The following will start a single cassandra node and run dbsetup.script
prior to executing tests,

  $ mvn clean test -Dcluster.mode=single

The following will start a 4 node cluster, run dbsetup.script, and set
the replication_factor for the keyspace to 2.

  $ mvn clean test -Dcluster.mode=multi -Dreplication.factor=2

This is also the initial commit for bundle.dbsetup.script which is
basically just a copy of dbsetup.script but is packaged in the cassandra
bundle. I wound up creating a separate file to avoid some complications
around filtering.